requirements.txt
greenlet==1.1.2
importlib-metadata==4.10.0
jellyfish==0.9.0
joblib==1.1.0
numpy==1.21.5
pandas==1.3.5
ply==3.11
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
recordtype==1.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.29
threadpoolctl==3.0.0
typing-extensions==4.0.1
zipp==3.7.0
